moved index.js test to separate file, ./test/index-test.js from ./test/fp-test.js and created README.md
