+++ lab-steve/README.md
@@ -1,0 +1,7 @@
+## Lab-Steve Lab-02 Documentation
+  * index.js calls functions 'main' function which returns CLI parameters as uppercase words using the fp.splice method from ./lib/fp.js.
+  * ./lib/fp.js defines the map, filter, reduce, concat, and splice methods using both call and apply.
+  * Tests
+    * ./test/fp-test.js tests the above methods
+    * **Stretch Goal:** ./test/index-test.js tests the index.js CLI functionality.
+  * Project passes Mocha tests and esLint.
